feat(client): enhance client app with new components and dependencies
- Added new components for creating and viewing experiences.
- Integrated Vanta.js for dynamic backgrounds based on user emotions.
- Updated dependencies including Radix UI, Tailwind CSS, and GSAP for improved UI/UX.
- Introduced new utility functions and UI components for better styling and functionality.
- Modified TypeScript configurations for improved path resolution.
- Updated Prisma schema to allow optional user IDs in the Petal model.
- Enhanced API endpoints for managing petals with create, read, update, and delete functionalities.

+++ apps/client/src/App.tsx
@@ -4,6 +4,16 @@ import gsap from "gsap";
 import AnimatedGradientBackground from "./Components/HomeBackground";
 import { CustomCursor } from "./Components/CustomCursor";
 import "./App.css";
+import { Button } from "./Components/components/ui/button";
+import { Flower, Plus } from "lucide-react";
+
+interface Petal {
+  id: number;
+  text: string;
+  dayOfWeek: string;
+  timeOfDay: string;
+  desiredEmotion: string;
+}
 
 function App() {
   const navigate = useNavigate();
@@ -13,14 +23,14 @@ function App() {
   const circleRef = useRef<HTMLDivElement>(null);
   const h1Ref = useRef<HTMLDivElement>(null);
   const svgPlusRef = useRef<HTMLDivElement | null>(null);
-  const svgRhombusRef = useRef<SVGSVGElement | null>(null);
   const petalsContainerRef = useRef<HTMLDivElement>(null);
   const petalsRef = useRef<(HTMLDivElement | null)[]>([]);
   const mainContentRef = useRef<HTMLDivElement>(null);
   const rotationAnimation = useRef<gsap.core.Tween | null>(null);
+  const [animationStarted, setAnimationStarted] = useState(false);
 
-  const [petalsData, setPetalsData] = useState<any[]>([]);
-  const [baseURL, setBaseUrl] = useState(
+  const [petalsData, setPetalsData] = useState<Petal[]>([]);
+  const [baseURL] = useState(
     `http://localhost:${import.meta.env.VITE_BACKEND_PORT || "8080"}`
   );
 
@@ -34,34 +44,51 @@ function App() {
   );
 
   useEffect(() => {
-    const pulse = gsap.to(circleRef.current, {
-      scale: 1.3,
-      duration: 1.2,
-      ease: "power1.inOut",
-      repeat: -1,
-      yoyo: true,
-    });
+    const circle = circleRef.current;
+    const pulseScale = 1.3;
+    const pulseDuration = 1.2;
+    const transitionDuration = 2;
 
-    const transitionDuration = 0.5;
+    if (circle && !animationStarted) {
+      setAnimationStarted(true);
+      gsap.fromTo(
+        circle,
+        { scale: 50 },
+        {
+          scale: 1,
+          duration: transitionDuration,
+          ease: "power2.out",
+          onComplete: () => {
+            const pulse = gsap.to(circle, {
+              scale: pulseScale,
+              duration: pulseDuration,
+              ease: "power1.inOut",
+              repeat: -1,
+              yoyo: true,
+            });
 
-    const time = setTimeout(() => {
-      gsap.to(circleRef.current, {
-        scale: 1,
-        duration: transitionDuration,
-        ease: "power2.out",
-        onComplete: () => {
-          pulse.kill();
-          setShowSplash(false);
-          setAnimationStart(true);
-          setIsLoading(false);
-        },
-      });
-    }, 7000);
+            const time = setTimeout(() => {
+              gsap.to(circle, {
+                scale: 1,
+                duration: 0.5,
+                ease: "power2.out",
+                onComplete: () => {
+                  pulse.kill();
+                  setShowSplash(false);
+                  setAnimationStart(true);
+                  setIsLoading(false);
+                },
+              });
+            }, 2000);
 
-    return () => {
-      clearTimeout(time);
-      pulse.kill();
-    };
+            return () => {
+              clearTimeout(time);
+              pulse.kill();
+            };
+          },
+        }
+      );
+    }
   }, []);
 
   useEffect(() => {
@@ -85,7 +112,7 @@ function App() {
         transformOrigin: "center center",
       }) as gsap.core.Tween;
 
-      const radius = 60;
+      const radius = 40;
       const centerX = 0;
       const centerY = 0;
       const deployDuration = 3;
@@ -105,7 +132,7 @@ function App() {
               y: centerY,
               scale: 0.5,
               opacity: 0,
-              transformOrigin: "0 0",
+              transformOrigin: "0 50%",
             },
             {
               duration: deployDuration,
@@ -122,27 +149,26 @@ function App() {
                   ease: "linear",
                   x: finalX,
                   y: finalY,
-                  transformOrigin: "0 0",
+                  transformOrigin: "0 50%",
                 });
               },
             }
           );
 
-          const helloSpan = petal.querySelector("span");
-          if (helloSpan) {
-            helloSpan.addEventListener("mouseenter", () => {
-              if (
-                rotationAnimation.current &&
-                !rotationAnimation.current.paused()
-              ) {
+          const petalSpan = petal.querySelector("span");
+          if (petalSpan) {
+            gsap.set(petalSpan, {
+              rotation: 2,
+              transformOrigin: "left center",
+            });
+
+            petalSpan.addEventListener("mouseenter", () => {
+              if (rotationAnimation.current) {
                 rotationAnimation.current.pause();
               }
             });
-            helloSpan.addEventListener("mouseleave", () => {
-              if (
-                rotationAnimation.current &&
-                rotationAnimation.current.paused()
-              ) {
+            petalSpan.addEventListener("mouseleave", () => {
+              if (rotationAnimation.current) {
                 rotationAnimation.current.resume();
               }
             });
@@ -179,6 +205,46 @@ function App() {
     }
   };
 
+  const handleGoToExperience = (petal: Petal) => {
+    const circle = circleRef.current;
+    const mainContent = mainContentRef.current;
+
+    if (circle) {
+      gsap.to(circle, {
+        scale: 50,
+        duration: 2,
+        borderRadius: 0,
+        zIndex: 10,
+        backgroundSize: "100% 100%",
+        backgroundPosition: "center center",
+        onComplete: () => {
+          gsap.set(mainContent, { opacity: 0, display: "none" });
+          navigate(`/experience/${petal.id}`);
+        },
+      });
+    }
+  };
+
+  const handleGoToCreate = () => {
+    const circle = circleRef.current;
+    const mainContent = mainContentRef.current;
+
+    if (circle) {
+      gsap.to(circle, {
+        scale: 50,
+        duration: 2,
+        borderRadius: 0,
+        zIndex: 10,
+        backgroundSize: "100% 100%",
+        backgroundPosition: "center center",
+        onComplete: () => {
+          gsap.set(mainContent, { opacity: 0, display: "none" });
+          navigate("/create");
+        },
+      });
+    }
+  };
+
   useEffect(() => {
     const fetchPetals = async () => {
       try {
@@ -205,7 +271,11 @@ function App() {
         className="flex flex-col items-center justify-between h-screen py-6 absolute top-0 left-0 w-full"
         style={{ zIndex: 5, opacity: 1 }}
       >
-        <h1 ref={h1Ref} className="" style={{ opacity: showSplash ? 0 : 1 }}>
+        <h1
+          ref={h1Ref}
+          className="border-1 border-black py-1 px-5 rounded-full"
+          style={{ opacity: showSplash ? 0 : 1 }}
+        >
           DAISY
         </h1>
         <div className="relative w-[100px] h-[100px] flex items-center justify-center">
@@ -215,9 +285,23 @@ function App() {
             style={{
               background: `linear-gradient(45deg, ${gradientColorsAmarillo.join(", ")})`,
               backgroundSize: `200% 200%`,
+              top: "50%",
+              left: "50%",
+              transform: "translate(-50%, -50%)",
             }}
           />
-          <div ref={petalsContainerRef} style={{ opacity: showSplash ? 0 : 1 }}>
+          <div
+            ref={petalsContainerRef}
+            className="absolute"
+            style={{
+              top: "50%",
+              left: "50%",
+              transform: "translate(-50%, -50%)",
+              opacity: showSplash ? 0 : 1,
+              width: "0px",
+              height: "0px",
+            }}
+          >
             {petalsData.map((petal, i) => {
               const angle = i * (360 / (petalsData.length || 6));
               return (
@@ -226,7 +310,7 @@ function App() {
                   className="absolute w-[30px] h-[30px] flex items-center justify-center"
                   style={{
                     transform: `rotate(${angle}deg)`,
-                    transformOrigin: "0 50%",
+                    transformOrigin: "center center",
                   }}
                   ref={(el) => {
                     if (el) {
@@ -235,12 +319,13 @@ function App() {
                   }}
                 >
                   <span
+                    onClick={() => handleGoToExperience(petal)}
                     style={{
-                      transform: `rotate(11deg) translateX(45px)`,
+                      transform: `rotate(1deg) translateX(45px)`,
                       whiteSpace: "nowrap",
                       fontSize: "1rem",
                     }}
-                    className="opacity-50 hover:opacity-100 transition-opacity duration-300 border-1 py-1 px-5 rounded-full petals-span"
+                    className="opacity-50 hover:opacity-100 transition-opacity duration-300 border-1 border-black py-1 px-5 rounded-full petals-span"
                   >
                     {petal.text || "Empty slot"}
                   </span>
@@ -250,55 +335,12 @@ function App() {
           </div>
         </div>
         <div
-          className="flex gap-2"
+          className="flex gap-1"
           ref={svgPlusRef}
           style={{ opacity: showSplash ? 0 : 1 }}
         >
-          <div>
-            <svg
-              width="36"
-              height="36"
-              viewBox="0 0 24 24"
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-              className="border-1 p-2 rounded-full svg-plus"
-            >
-              <path
-                d="M12 4V20M4 12H20"
-                stroke="currentColor"
-                strokeWidth="2"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-              />
-            </svg>
-          </div>
-          <div>
-            <svg
-              width="36"
-              height="36"
-              viewBox="0 0 36 36"
-              xmlns="http://www.w3.org/2000/svg"
-              className="border-1 p-2 rounded-full svg-plus"
-              ref={svgRhombusRef}
-              onClick={handleGoToList}
-            >
-              <circle
-                cx="18"
-                cy="18"
-                r="16"
-                stroke="currentColor"
-                strokeWidth="2"
-              />
-              <path
-                d="M18 4L32 18L18 32L4 18L18 4Z"
-                fill="currentColor"
-                stroke="currentColor"
-                strokeWidth="2"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-              />
-            </svg>
-          </div>
+          <Button onClick={handleGoToCreate} icon={<Plus />} />
+          <Button onClick={handleGoToList} icon={<Flower />} />
         </div>
       </div>
     </div>
